requirements.txt
openai>=1.0.0
python-dotenv==1.0.0
Flask==3.0.2
flask-cors==4.0.0
gunicorn==21.2.0
